sync web generators
