--- web/lib/generator/dart_class.dart
@@ -1,9 +1,0 @@
-class DartClass {
-  final String className;
-  final String classCode;
-
-  DartClass({
-    required this.className,
-    required this.classCode,
-  });
-}
