--- web/lib/generator/models.dart
@@ -1,16 +1,0 @@
-import 'utils.dart';
-import 'package:yaml/yaml.dart';
-
-import 'swagger.dart';
-
-String generateModels(DatabaseSwagger swagger, YamlMap? mappings) {
-  String code = 'library models;\n';
-  code += "export 'client_extension.dart';\n";
-  code += "export 'supadart_abstract_class.dart';\n";
-  swagger.definitions.forEach((tableName, _) {
-    final className = tableNameToClassName(tableName, mappings);
-    final fileName = classNameToFileName(className);
-    code += "export '$fileName';\n";
-  });
-  return code;
-}
