--- web/lib/generator/static_column_names.dart
@@ -1,9 +1,0 @@
-import 'swagger.dart';
-
-String generateStaticColumnNames(Map<String, Column> properties) {
-  String code = "";
-  properties.forEach((propertyName, col) {
-    code += "static String get c_$propertyName => '${col.dbColName}';\n";
-  });
-  return "$code\n";
-}
