+++ web/lib/generators/class/class.dart
