+++ web/lib/generators/utils/fetch_swagger.dart
